# shared/api/bundles/order.openapi.v1.bundle.yaml
openapi: 3.0.3
info:
  title: OrderService API
  version: 1.0.0
  description: API для работы с заказами
tags:
  - name: Order
    description: Операции с заказами и обновление информации
paths:
  /api/v1/orders:
    post:
      summary: Создаёт новый заказ на основе выбранных пользователем деталей.
      operationId: CreateNewOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_order_request'
      responses:
        '201':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/create_order_response'
        '400':
          description: Одна или несколько деталей не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
  /api/v1/orders/{order_uuid}/pay:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      summary: Проводит оплату ранее созданного заказа.
      operationId: OrderPay
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pay_order_request'
      responses:
        '200':
          description: Заказ успешно оплачен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pay_order_response'
        '404':
          description: Заказ не существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
  /api/v1/orders/{order_uuid}:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    get:
      summary: Возвращает информацию о заказе.
      operationId: GetOrderByUuid
      tags:
        - Order
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order_dto'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
  /api/v1/orders/{order_uuid}/cancel:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      summary: Отменяет заказ.
      operationId: CancelOrder
      tags:
        - Order
      responses:
        '204':
          description: Заказ успешно удален
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: заказ уже оплачен и не может быть отменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
components:
  schemas:
    create_order_request:
      type: object
      required:
        - user_uuid
        - part_uuids
      properties:
        user_uuid:
          type: string
          format: uuid
          description: UUID пользователя
        part_uuids:
          type: array
          items:
            type: string
            format: uuid
          description: Список UUID деталей
    create_order_response:
      type: object
      required:
        - uuid
        - total_price
      properties:
        uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор заказа
        total_price:
          type: number
          format: double
          description: Итоговая стоимость
    bad_request_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 400
        message:
          type: string
          description: Описание ошибки
          example: Некорректный запрос
    payment_method:
      type: string
      description: Способ оплаты (если оплачен)
      enum:
        - PAYMENT_METHOD_UNKNOWN
        - PAYMENT_METHOD_CARD
        - PAYMENT_METHOD_SBP
        - PAYMENT_METHOD_CREDIT_CARD
        - PAYMENT_METHOD_INVESTOR_MONEY
      x-enum-descriptions:
        - Неизвестный способ
        - Банковская карта
        - Система быстрых платежей
        - Кредитная карта
        - Деньги инвестора
    pay_order_request:
      type: object
      required:
        - payment_method
      properties:
        payment_method:
          $ref: '#/components/schemas/payment_method'
    pay_order_response:
      type: object
      required:
        - transaction_uuid
      properties:
        transaction_uuid:
          type: string
          format: uuid
          description: UUID транзакции
    not_found_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Описание ошибки
          example: Не найдено
    order_status:
      type: string
      description: Статус (PENDING_PAYMENT, PAID, CANCELLED)
      enum:
        - PENDING_PAYMENT
        - PAID
        - CANCELLED
    order_dto:
      type: object
      required:
        - order_uuid
        - user_uuid
        - part_uuids
        - total_price
        - status
      properties:
        order_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор заказа
        user_uuid:
          type: string
          format: uuid
          description: UUID пользователя
        part_uuids:
          type: array
          items:
            type: string
            format: uuid
          description: Список UUID деталей
        total_price:
          type: number
          format: double
          description: Итоговая стоимость
        transaction_uuid:
          type: string
          format: uuid
          description: UUID транзакции (если оплачен)
        payment_method:
          $ref: '#/components/schemas/payment_method'
        status:
          $ref: '#/components/schemas/order_status'
    conflict_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 409
        message:
          type: string
          description: Описание ошибки
          example: Конфликт
  parameters:
    order_uuid:
      name: order_uuid
      in: path
      required: true
      description: UUID заказа который надо оплатить
      schema:
        type: string
        format: uuid
x-ogen:
  target: ./shared/pkg/openapi/order/v1
  package: order_v1
  clean: true